refactor: 화면에 따른 패키지 이동
