+++ app/src/main/java/woowacourse/shopping/ui/productdetail/ProductDetailActivity.kt
@@ -1,10 +1,12 @@
-package woowacourse.shopping
+package woowacourse.shopping.ui.productdetail
 
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.viewModels
+import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityProductDetailBinding
+import woowacourse.shopping.ui.base.BaseActivity
 
 class ProductDetailActivity : BaseActivity<ActivityProductDetailBinding>(R.layout.activity_product_detail) {
     val viewModel: ProductDetailViewModel by viewModels()
