+++ app/src/main/java/woowacourse/shopping/ui/productdetail/ProductDetailViewModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping
+package woowacourse.shopping.ui.productdetail
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
