+++ app/src/main/res/layout/activity_product_detail.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="woowacourse.shopping.ProductDetailViewModel" />
+            type="woowacourse.shopping.ui.productdetail.ProductDetailViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
