+++ app/src/main/res/layout/item_product.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="onClickHandler"
-            type="woowacourse.shopping.ProductViewHolder.OnClickHandler" />
+            type="woowacourse.shopping.ui.products.ProductViewHolder.OnClickHandler" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
